Machine Learning/ChatBot/requirements.txt
flask==3.0.*
functions-framework==3.*
google-cloud-bigquery==3.23.0
google-cloud-storage==2.17.0
pandas==2.2.*
gunicorn